+++ R/utils.R
@@ -92,8 +92,8 @@ delete_confirmed <- function(confirm, name, type, quote_name=TRUE)
         return(TRUE)
 
     msg <- if(quote_name)
-        sprintf("Do you really want to delete the %s '%s'? (y/N) ", type, name)
-    else sprintf("Do you really want to delete the %s %s? (y/N) ", type, name)
-    yn <- readline(msg)
-    return(tolower(substr(yn, 1, 1)) == "y")
+        sprintf("Do you really want to delete the %s '%s'?", type, name)
+    else sprintf("Do you really want to delete the %s %s?", type, name)
+    ok <- utils::askYesNo(msg, FALSE)
+    return(isTRUE(ok))
 }